code review guidelines: enforce backwards-compatibility in migrations
Migrations should be backwards compatible. If there is a migration that
would be incompatible with existing code, users would experience downtime
during the migration or we'd need to use subdeployments.

To reduce the complexity of our tooling, splitting the change between
two deployments in a way that the migration never results in
a period of backwards incompatible code is the simplest solution for now.

There is a discussion in the original issue.

Fixes https://github.com/shuddle/engineering/issues/44